Move index into lib, minor tweaks to files for new transpile script.
