+++ lib/collection.js
@@ -12,6 +12,7 @@ const $client = Symbol('client');
 
 export class Collection extends Map {
   constructor(client) {
+    super();
     this[$client] = client;
     this[$filters] = [];
     this[$view] = View;
@@ -112,8 +113,8 @@ export class Collection extends Map {
     // todo: add to any view of which this matches the filter(?)
     super.set(model.identity(), model);
     this[$views].forEach(view => {
-      if (this)
-      
+      //if (this)
+
     });
     return this;
   }
