+++ lib/collection/filter.js
@@ -16,4 +16,4 @@ export class Filter {
   get name() {
     return this[$name];
   }
-}
+};
