+++ lib/view.js
@@ -10,6 +10,7 @@ const $collection = Symbol('collection');
 
 export default class View extends Map {
   constructor(promise, collection, params) {
+    super();
     this[$collection] = collection;
     this[$params] = params;
     this[$loaded] = false;
