+++ package.json
@@ -18,11 +18,13 @@
     "url": "https://github.com/epixa/jsclient/issues"
   },
   "homepage": "https://github.com/epixa/jsclient",
+  "main": "build/index.js",
   "dependencies": {
     "bluebird": "^2.9.26",
     "lodash": "^3.9.2"
   },
   "devDependencies": {
+    "babel": "^5.8.20",
     "babel-core": "^5.8.20",
     "babel-loader": "^5.3.2",
     "karma": "^0.13.3",
@@ -41,5 +43,6 @@
     "install": "npm run build",
     "build": "./node_modules/webpack/bin/webpack.js",
     "test": "./node_modules/karma/bin/karma start ./karma.config.js"
+    "transpile": "babel lib --out-dir build",
   }
 }
