+++ webpack.config.js
@@ -2,11 +2,10 @@ var webpack = require('webpack'),
   path = require('path');
 
 module.exports = {
-  entry: './index.js',
+  entry: './lib/index.js',
   output: {
     path: path.join(__dirname, '/dist'),
-    filename: "bundle.js",
-    sourceMapFilename: "bundle.map"
+    filename: "jsclient.js"
   },
   module: {
     loaders: [{
